refactor: move AflamiException to domain exceptions package
